Implement vertex retrieval for `ArrayTree` type
This commit implements vertex retrieval for the `ArrayTree` type, via
the `get_vertex` function.

To achieves this, a private helper of the same name and signature has
been implemented on the (also private) `ArrayTreeNode` type. This
helper is where most of the logic is implemented.

Furthermore, to aid in depth-first search, the vertex number (i.e.
`VertexNumber`) has been added to the `ArrayTreeNode` definition.